+++ CHANGELOG.md
@@ -1,3 +1,7 @@
-## 0.0.1
+## 1.0.0
 
-* TODO: Describe initial release.
+* Inherits all the functionalities of the standard Text widget.
+* Defines the minimum number of lines the widget will always display.
+* Allows users to select and highlight specific portions of the text.
+* Provides access to a set of pre-defined text styles commonly used in user interfaces.
+